Fix Collections Character and Film

+++ SpringSakilaInventory/src/main/java/com/springsakila/controllers/FilmController.java
@@ -47,7 +47,8 @@ public class FilmController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable int id) {
+    public void deleteById(@PathVariable int id) throws NotFoundException {
+        if(filmService.getOne(id).isEmpty()) throw new NotFoundException();
         filmService.deleteById(id);
     }
